feat: refactor source
THe src folder contains main.py and modules/*.py, containing all modules

BREAKING CHANGE: The build process has been revamped
