+++ .github/workflows/main.yml
@@ -54,66 +54,3 @@ jobs:
         with:
           name: "DiSH-${{ runner.os }}"
           path: "DiSH-*.zip"
-  deploy:
-    runs-on: ubuntu-latest
-    if: |
-      endsWith(github.event.head_commit.message, ' !') 
-    needs: ["build"]
-    steps:
-    
-      - name: Check out
-        uses: actions/checkout@v2
-        with:
-          token: "${{ secrets.GITHUB_TOKEN }}"
-          fetch-depth: 0
-
-      - name: Install additonal deps
-        if: runner.os == 'Linux'
-        run: sudo apt-get install -y python3-dev libasound2-dev
-      - name: Setup Python
-        uses: actions/setup-python@v4.2.0
-      - name: Setup packages
-        run: |
-          python -m pip install -r requirements.txt
-          python -m  pip install -U Commitizen  
-      - name: Bump and changelog
-        run: |
-          cz bump --files-only
-          cz changelog
-      - name: Install additonal deps
-        if: runner.os == 'Linux'
-        run: sudo apt-get install -y python3-dev libasound2-dev
-      - name: Git Commit and Push
-        uses: EndBug/add-and-commit@v8
-        with:
-          
-          message: "bump: update version"
-          author_name: github-actions[bot]
-          author_email: github-actions[bot]@users.noreply.github.com
-      - name: Download build artifact (Linux)
-        uses: actions/download-artifact@v3.0.0
-        with:
-          name: DiSH-Linux
-      - name: Download build artifact (Windows)
-        uses: actions/download-artifact@v3.0.0
-        with:
-          name: DiSH-Windows
-      - name: Print Version
-        run: echo "Bumped to version ${{ steps.cz.outputs.version }}"
-      
-      - name: Get short hash
-        run: |
-          echo "VER=$(python version_exporter.py)" >> $GITHUB_ENV
-          python version_exporter.py
-      - name: Create Relese
-        uses: ncipollo/release-action@v1.10.0
-        with:
-          artifacts: "DiSH-*"
-          tag: "v${{ env.VER }}"
-          name: "DiSH-${{ env.VER }}"
-          bodyFile: "CHANGELOG.md"
-     
-  
-      
-      
-
