--- .gitmodules
@@ -1,5 +1,0 @@
-
-[submodule "DiSHScript"]
-	path = DiSHScript
-	url = https://github.com/LDevs-Team/DiSHScript/
-	branch = main
