+++ src/modules/__init__.py
@@ -1,0 +1,1 @@
+from . import camera, filesystem, keyboard, mouse, misc, sound, screen
