+++ src/modules/camera.py
@@ -1,0 +1,17 @@
+
+import os
+import tempfile
+import cv2
+import discord
+
+
+async def cam(client:discord.Client, message:discord.Message, args:str):
+    with tempfile.NamedTemporaryFile(suffix=".png", delete=False) as temp:
+        o = cv2.VideoCapture(0)
+        s, img = o.read()
+        del s
+        cv2.imwrite(temp.name, img)
+        await message.channel.send(file=discord.File(temp.name))
+        o.release()
+        temp.close()
+        os.unlink(temp.name)
